Refactor notification scheduling
