+++ OnePushup/Services/INotificationScheduler.cs
@@ -1,0 +1,9 @@
+namespace OnePushUp.Services;
+
+public interface INotificationScheduler
+{
+    Task ScheduleAsync(TimeSpan time);
+    Task CancelAsync();
+    Task SendTestAsync();
+}
+
